Moving fixture to separate file
