+++ cfme/tests/infrastructure/test_webmks_console.py
@@ -40,12 +40,6 @@ def vm_obj(request, provider, setup_provider, console_template, vm_name):
         logger.warning("Failed to delete vm `{}`.".format(vm_obj.name))
 
 
-@pytest.fixture(scope="module")
-def configure_vmware_console_for_test(appliance):
-    """Configure VMware Console to use VNC which is what is required for the HTML5 console."""
-    appliance.server.settings.update_vmware_console({'console_type': 'VMware WebMKS'})
-
-
 @pytest.yield_fixture
 def ssh_client(vm_obj, console_template):
     """Provide vm_ssh_client for ssh operations in the test."""
@@ -56,26 +50,8 @@ def ssh_client(vm_obj, console_template):
         yield vm_ssh_client
 
 
-@pytest.fixture(scope="module")
-def configure_websocket(appliance):
-    """
-    Enable websocket role if it is disabled.
-
-    Currently the fixture cfme/fixtures/base.py,
-    disables the websocket role to avoid intrusive popups.
-    """
-    server_settings = appliance.server.settings
-    roles = server_settings.server_roles_db
-    if 'websocket' in roles and not roles['websocket']:
-        logger.info('Enabling the websocket role to allow console connections')
-        server_settings.enable_server_roles('websocket')
-        yield
-    logger.info('Disabling the websocket role to avoid intrusive popups')
-    server_settings.disable_server_roles('websocket')
-
-
 def test_webmks_vm_console(request, appliance, provider, vm_obj, configure_websocket,
-        configure_vmware_console_for_test, take_screenshot, ssh_client):
+        configure_console_webmks, take_screenshot, ssh_client):
     """Test the VMware WebMKS console support for a particular provider.
 
     The supported providers are:
